async_sock: Removed continous loop
Async ZMQ sockets no longer run in a continous loop to check for
messages. I've essentially implemented a loop within asyncio loop,
but there is a major flaw - number of sockets is limited by the number
of threads in the executor of asyncio loop. It is possible to add your
own executor, but it is not a clean solution. This is an intermediary
stage, which will be changed in the next revision.