async_sock: Removed socket to thread dependency
Prior major flaw, which used a single thread for each socket to poll for
socket events was removed. Essentially this implements a full loop,
which is integrated into asyncio loop. Main difference - a single thread
is used by async_zmq to poll for all sockets that could be created with
SocketFactory.